Move the Dusk macros to a mixin class to be used in userland apps too
