+++ tests/Browser/TestCase.php
@@ -7,15 +7,14 @@ use Exception;
 use Psy\Shell;
 use Throwable;
 use Laravel\Dusk\Browser;
-use Illuminate\Support\Str;
 use Illuminate\Support\Facades\File;
 use Illuminate\Support\Facades\Route;
 use Livewire\LivewireServiceProvider;
 use Illuminate\Support\Facades\Artisan;
-use PHPUnit\Framework\Assert as PHPUnit;
 use Facebook\WebDriver\Chrome\ChromeOptions;
 use Facebook\WebDriver\Remote\RemoteWebDriver;
 use Facebook\WebDriver\Remote\DesiredCapabilities;
+use Livewire\Macros\DuskBrowserMacros;
 use Orchestra\Testbench\Dusk\Options as DuskOptions;
 use Orchestra\Testbench\Dusk\TestCase as BaseTestCase;
 
@@ -28,7 +27,7 @@ class TestCase extends BaseTestCase
             DuskOptions::withoutUI();
         }
 
-        $this->registerMacros();
+        Browser::mixin(DuskBrowserMacros::class);
 
         $this->afterApplicationCreated(function () {
             $this->makeACleanSlate();
@@ -169,128 +168,6 @@ class TestCase extends BaseTestCase
     protected function livewireViewsPath($path = '')
     {
         return resource_path('views').'/livewire'.($path ? '/'.$path : '');
-    }
-
-    protected function registerMacros()
-    {
-        Browser::macro('assertAttributeMissing', function ($selector, $attribute) {
-            $fullSelector = $this->resolver->format($selector);
-
-            $actual = $this->resolver->findOrFail($selector)->getAttribute($attribute);
-
-            PHPUnit::assertNull(
-                $actual,
-                "Did not see expected attribute [{$attribute}] within element [{$fullSelector}]."
-            );
-
-            return $this;
-        });
-
-        Browser::macro('assertNotVisible', function ($selector) {
-            $fullSelector = $this->resolver->format($selector);
-
-            PHPUnit::assertFalse(
-                $this->resolver->findOrFail($selector)->isDisplayed(),
-                "Element [{$fullSelector}] is visible."
-            );
-
-            return $this;
-        });
-
-        Browser::macro('assertNotPresent', function ($selector) {
-            $fullSelector = $this->resolver->format($selector);
-
-            PHPUnit::assertTrue(
-                is_null($this->resolver->find($selector)),
-                "Element [{$fullSelector}] is present."
-            );
-
-            return $this;
-        });
-
-        Browser::macro('assertHasClass', function ($selector, $className) {
-            /** @var \Laravel\Dusk\Browser $this */
-            $fullSelector = $this->resolver->format($selector);
-
-            PHPUnit::assertContains(
-                $className,
-                explode(' ', $this->attribute($selector, 'class')),
-                "Element [{$fullSelector}] missing class [{$className}]."
-            );
-
-            return $this;
-        });
-
-        Browser::macro('assertScript', function ($js, $expects = true) {
-            PHPUnit::assertEquals($expects, head($this->script(
-                Str::start( $js, 'return ')
-            )));
-
-            return $this;
-        });
-
-        Browser::macro('assertClassMissing', function ($selector, $className) {
-            /** @var \Laravel\Dusk\Browser $this */
-            $fullSelector = $this->resolver->format($selector);
-
-            PHPUnit::assertNotContains(
-                $className,
-                explode(' ', $this->attribute($selector, 'class')),
-                "Element [{$fullSelector}] has class [{$className}]."
-            );
-
-            return $this;
-        });
-
-        Browser::macro('waitForLivewireToLoad', function () {
-            return $this->waitUsing(5, 75, function () {
-                return $this->driver->executeScript("return !! window.Livewire.components.initialRenderIsFinished");
-            });
-        });
-
-        Browser::macro('waitForLivewire', function ($callback = null) {
-            $id = rand(100, 1000);
-
-            $this->script([
-                "window.duskIsWaitingForLivewireRequest{$id} = true",
-                "window.Livewire.hook('message.sent', () => { window.duskIsWaitingForLivewireRequest{$id} = true })",
-                "window.Livewire.hook('message.processed', () => { delete window.duskIsWaitingForLivewireRequest{$id} })",
-                "window.Livewire.hook('message.failed', () => { delete window.duskIsWaitingForLivewireRequest{$id} })",
-            ]);
-
-            if ($callback) {
-                $callback($this);
-
-                // Wait a quick sec for Livewire to hear a click and send a request.
-                $this->pause(25);
-
-                return $this->waitUsing(5, 50, function () use ($id) {
-                    return $this->driver->executeScript("return window.duskIsWaitingForLivewireRequest{$id} === undefined");
-                }, 'Livewire request was never triggered');
-            }
-
-            // If no callback is passed, make ->waitForLivewire a higher-order method.
-            return new class($this, $id) {
-                public function __construct($browser, $id) { $this->browser = $browser; $this->id = $id; }
-
-                public function __call($method, $params)
-                {
-                    return tap($this->browser->{$method}(...$params), function ($browser) {
-                        $browser->waitUsing(5, 25, function () use ($browser) {
-                            return $browser->driver->executeScript("return window.duskIsWaitingForLivewireRequest{$this->id} === undefined");
-                        }, 'Livewire request was never triggered');
-                    });
-                }
-            };
-        });
-
-        Browser::macro('online', function () {
-            return tap($this)->script("window.dispatchEvent(new Event('online'))");
-        });
-
-        Browser::macro('offline', function () {
-            return tap($this)->script("window.dispatchEvent(new Event('offline'))");
-        });
     }
 
     protected function driver(): RemoteWebDriver
